+++ packages/terra-framework-docs/src/terra-dev-site/doc/tabs/example/IconsInMenuTabs.jsx
@@ -14,45 +14,45 @@ const cx = classNames.bind(styles);
 
 const IconsOnlyAndInMenuTabs = () => {
   const searchTab = (
-    <Tabs.Pane label="Search" icon={<IconSearch />} showIcon key="Search">
-      <TabContent label="Search" />
+    <Tabs.Pane label="Search" icon={<IconSearch />} showIcon key="MenuTabSearch">
+      <TabContent label="Search" id="MenuTabSearchContent" />
     </Tabs.Pane>
   );
 
   const briefcaseTab = (
     // eslint-disable-next-line react/forbid-component-props
-    <Tabs.Pane label="Briefcase" icon={<IconBriefcase />} showIcon key="Briefcase">
-      <TabContent label="Briefcase" />
+    <Tabs.Pane label="Briefcase" icon={<IconBriefcase />} showIcon key="MenuTabBriefcase">
+      <TabContent label="Briefcase" id="MenuTabBriefcaseContent" />
     </Tabs.Pane>
   );
 
   const bookmarkTab = (
-    <Tabs.Pane label="Bookmark" icon={<IconBookmark />} showIcon key="Bookmark">
-      <TabContent label="Bookmark" />
+    <Tabs.Pane label="Bookmark" icon={<IconBookmark />} showIcon key="MenuTabBookmark">
+      <TabContent label="Bookmark" id="MenuTabBookmarkContent" />
     </Tabs.Pane>
   );
 
   const calendarTab = (
-    <Tabs.Pane label="Calendar" icon={<IconCalendar />} showIcon key="Calendar">
-      <TabContent label="Calendar" />
+    <Tabs.Pane label="Calendar" icon={<IconCalendar />} showIcon key="MenuTabCalendar">
+      <TabContent label="Calendar" id="MenuTabCalendarContent" />
     </Tabs.Pane>
   );
 
   const envelopeTab = (
-    <Tabs.Pane label="Envelope" icon={<IconEnvelope />} showIcon key="Envelope">
-      <TabContent label="Envelope" />
+    <Tabs.Pane label="Envelope" icon={<IconEnvelope />} showIcon key="MenuTabEnvelope">
+      <TabContent label="Envelope" id="MenuTabEnvelopeContent" />
     </Tabs.Pane>
   );
 
   const printerTab = (
-    <Tabs.Pane label="Printer" icon={<IconPrinter />} showIcon key="Printer">
-      <TabContent label="Printer" />
+    <Tabs.Pane label="Printer" icon={<IconPrinter />} showIcon key="MenuTabPrinter">
+      <TabContent label="Printer" id="MenuTabPrinterContent" />
     </Tabs.Pane>
   );
 
   return (
     <div className={cx('content-wrapper')}>
-      <Tabs setFocusOnContent id="icons-only">
+      <Tabs setFocusOnContent id="icons-only-menu">
         {searchTab}
         {briefcaseTab}
         {bookmarkTab}