+++ packages/terra-table/src/Table.jsx
@@ -90,6 +90,7 @@ const propTypes = {
 
   /**
    * Callback function that is called when a selectable cell is selected. Parameters:
+   * @private
    * @param {string} rowId rowId
    * @param {string} columnId columnId
    */
@@ -111,6 +112,11 @@ const propTypes = {
    * rendered to allow for row selection to occur.
    */
   hasSelectableRows: PropTypes.bool,
+
+  /**
+   * Boolean specifying whether or not the table should have zebra striping for rows.
+   */
+  isStriped: PropTypes.bool,
 };
 
 const defaultProps = {
@@ -143,6 +149,7 @@ function Table(props) {
     onCellSelect,
     onSectionSelect,
     hasSelectableRows,
+    isStriped,
     rowHeaderIndex,
   } = props;
 
@@ -160,9 +167,10 @@ function Table(props) {
   const activeColumnPageX = useRef(0);
   const activeColumnWidth = useRef(200);
   const tableWidth = useRef(0);
-  const tableRef = useRef();
 
-  // const sortedColumn = useRef();
+  const tableContainerRef = useRef();
+  const tableRef = useRef();
+  const [isTableScrollable, setTableScrollable] = useState(false);
 
   const gridContext = useContext(GridContext);
   const theme = useContext(ThemeContext);
@@ -233,6 +241,10 @@ function Table(props) {
   useEffect(() => {
     const resizeObserver = new ResizeObserver(() => {
       setTableHeight(tableRef.current.offsetHeight - 1);
+
+      const tableContainer = tableContainerRef.current;
+      setTableScrollable(tableContainer.scrollWidth > tableContainer.clientWidth
+                        || tableContainer.scrollHeight > tableContainer.clientHeight);
     });
 
     resizeObserver.observe(tableRef.current);
@@ -293,7 +305,12 @@ function Table(props) {
   // -------------------------------------
 
   return (
-    <div className={cx('table-container')}>
+    <div
+      ref={tableContainerRef}
+      className={cx('table-container')}
+      // eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex
+      tabIndex={!isGridContext && isTableScrollable ? 0 : undefined}
+    >
       <table
         ref={tableRef}
         id={id}