+++ packages/terra-table/src/proptypes/cellShape.js
@@ -6,14 +6,16 @@ const cellShape = PropTypes.shape({
    */
   content: PropTypes.node,
   /**
-   *  Boolean indicating if cell contents are masked.
+   * Boolean indicating if cell contents are masked.
    */
   isMasked: PropTypes.bool,
   /**
+   * @private
    * Boolean value indicating whether or not the column header is selectable.
    */
   isSelectable: PropTypes.bool,
   /**
+   * @private
    * Boolean value indicating whether or not the cell should render as selected.
    */
   isSelected: PropTypes.bool,