+++ packages/terra-file-path/CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## Unreleased
 
+## 0.3.0 - (October 17, 2023)
+
 * Inital stable release.