+++ packages/terra-table/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+* Added
+  * Added the ability to toggle zebra striping for table rows.
+  
+* Changed
+  * Updated the table component so that the cell dive-in logic would not execute when not in the grid context.
+  * Modified the table component so that it can receive focus when scrollable.
+
+## 5.0.0-alpha.0 - (October 17, 2023)
+
 * Breaking Change
   * Redesigned the table component to meet accessibility guidelines. See Docs for details of the new API.